// typechain-types/GuardedExecutor.ts
export declare namespace GuardedExecutor {
  export type RequestStruct = {
    id: BytesLike;
    policyId: BytesLike;
    submitter: AddressLike;
    recipient: AddressLike;
    amount: BigNumberish;
    token: AddressLike;
    evidenceHash: BytesLike;
    storageRef: string;
    memo: string;
    status: BigNumberish;
    receiptId: BytesLike;
    createdAt: BigNumberish;
    resolvedAt: BigNumberish;
  };

  export type RequestStructOutput = [
    id: string,
    policyId: string,
    submitter: string,
    recipient: string,
    amount: bigint,
    token: string,
    evidenceHash: string,
    storageRef: string,
    memo: string,
    status: bigint,
    receiptId: string,
    createdAt: bigint,
    resolvedAt: bigint
  ] & {
    id: string;
    policyId: string;
    submitter: string;
    recipient: string;
    amount: bigint;
    token: string;
    evidenceHash: string;
    storageRef: string;
    memo: string;
    status: bigint;
    receiptId: string;
    createdAt: bigint;
    resolvedAt: bigint;
  };
}