// typechain-types/ReceiptRegistry.ts
export declare namespace ReceiptRegistry {
  export type ReceiptStruct = {
    id: BytesLike;
    requestId: BytesLike;
    evidenceHash: BytesLike;
    txHash: BytesLike;
    status: BigNumberish;
    executor: AddressLike;
    submitter: AddressLike;
    amount: BigNumberish;
    token: AddressLike;
    recipient: AddressLike;
    createdAt: BigNumberish;
    confirmedAt: BigNumberish;
    storageRef: string;
    failReason: string;
  };

  export type ReceiptStructOutput = [
    id: string,
    requestId: string,
    evidenceHash: string,
    txHash: string,
    status: bigint,
    executor: string,
    submitter: string,
    amount: bigint,
    token: string,
    recipient: string,
    createdAt: bigint,
    confirmedAt: bigint,
    storageRef: string,
    failReason: string
  ] & {
    id: string;
    requestId: string;
    evidenceHash: string;
    txHash: string;
    status: bigint;
    executor: string;
    submitter: string;
    amount: bigint;
    token: string;
    recipient: string;
    createdAt: bigint;
    confirmedAt: bigint;
    storageRef: string;
    failReason: string;
  };
}